Create upstream.go
Create upstream.go and move all code related to the upstream proxy
feature.
